+++ src/pages/Pool/index.tsx
@@ -14,7 +14,7 @@ import { ThemedText } from 'theme/components'
 import { ButtonRow, ErrorContainer, LoadingRows, MainContentWrapper, NetworkIcon, OnlyRewardedSwitcher, OnlyRewardedSwitcherContainer, OnlyRewardedSwitcherLabel, PageHeader, PageWrapper, Panel, PanelTopLight, PanelWrapper, ResponsiveButtonPrimary, ResponsiveButtonTabs, TitleRow } from './styled'
 import { getAllPools } from 'api/PoolsData'
 import Pools from 'components/Pools'
-import { formattedNum, formattedPercent, get2DayPercentChange, getPercentChange } from 'utils/dashboard'
+import { formattedNum, formattedPercent, get2DayPercentChange, getPercentChange } from 'utils/formatNum'
 import { REWARDS_SELECTOR, STARKNET_REWARDS_API_URL } from 'constants/strkRewards'
 import { HISTORICAL_GLOBAL_DATA } from 'apollo/queries'
 import { apiTimeframeOptions } from 'constants/apiTimeframeOptions'
